requirements.txt
asgiref==3.3.0
autopep8==1.5.4
Django==3.1.3
djangorestframework==3.12.2
joblib==0.17.0
numpy==1.19.4
pandas==1.1.4
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.22.1
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2